Fix compile warnings: -Wsign-conversion in InstrumentTrackState/Compressor/AlsaAudio
Stage 3 of plans/warning-cleanup.md, continuing largest-first.
Mechanical index/argument casts throughout - no behavior changes.

Clean build: 56 -> 32 warnings, all -Wsign-conversion remaining.

Verified: clean rebuild (0 errors), full ctest pass.

+++ InstrumentTrackState.h
@@ -30,7 +30,7 @@ public:
     std::vector<std::pair<int, AudioBuffer> > chunks;
 
     if (instrument_id_ >= 0 && instrument_id_ < static_cast<int>(instruments.size())) {
-      auto & instrument = instruments[instrument_id_];
+      auto & instrument = instruments[static_cast<size_t>(instrument_id_)];
       auto & pending_events = context.getPendingEvents(track_id_);
       auto & pending_azimuth = context.getPendingAzimuthTicks(track_id_);
